fixing issue that truncates times on redhat

+++ src/objects/clock.c
@@ -107,6 +107,6 @@ void CLOCK_pTicks(CLOCK*   cl,
 /* Convert a time in ticks (time_t) to msecs (float) */
 float ticks_to_msec(time_t t)
 {
-   return ((float)t * 1000.f) / ((float)CLOCKS_PER_SEC);
+   return ( (float)t * (float)1000.f ) / ( (float)CLOCKS_PER_SEC );
 }
